+++ libs/core/src/exception_unittest.cpp
@@ -52,11 +52,15 @@ TEST(exception, stackedExceptionComplex)
 	{
 		const auto sExc = mrpt::exception_to_str(e);
 		const auto num_lines = std::count(sExc.begin(), sExc.end(), '\n');
-		EXPECT_GT(num_lines, 10);
+		EXPECT_GT(num_lines, 10) << sExc;
 
 		EXPECT_TRUE(sExc.find("Message:  Aw!") != std::string::npos) << sExc;
+
+// This test doesn't pass in Windows if building w/o debug symbols:
+#if !defined(_WIN32) || defined(_DEBUG)
 		EXPECT_TRUE(sExc.find("test_except_toplevel") != std::string::npos)
 			<< sExc;
+#endif
 	}
 }